Add files via upload
The notebook file is essentially the space where I developed & test each step of the registration of light-sheet to phase data. It uses timepoint 0 of each dataset from the latest neuromast dataset acquired. The two script files basically split the notebook into a function file and an application file, but the documentation isn't finished (sorry). The script needs to be adapted to pull info directly from zarr stores and then write results into a new one. Hopefully that goes smoothly. I will add that improvement to this PR ASAP. The other thing that needs to be added is the ICP-based algorithm that I use for bead matching between data from the two arms. This is essential. Will also upload that ASAP.